DASH-1191 Move tests in separate files and reuse them for testnet and local dev net.
