+++ .env.test.example
@@ -6,6 +6,10 @@ PUBLIC_KEY=
 PRIVATE_KEY_2=
 PUBLIC_KEY_2=
 
+# Faucet
+FAUCET_PRIVATE_KEY=
+FAUCET_PUBLIC_KEY=
+
 # FIO Address of Wallet 1
 TEST_FIO_ADDRESS_NAME=
 
@@ -15,49 +19,6 @@ TEST_FIO_ADDRESS_NAME_2=
 # FIO Domain of Wallet 1
 TEST_FIO_DOMAIN_NAME=
 
-# Encrypted Wallet 1
-ENC_PRIVATE_KEY=
-ENC_PUBLIC_KEY=
-
-# Encrypted Wallet 2
-ENC_PRIVATE_KEY_2=
-ENC_PUBLIC_KEY_2=
-
-# FIO Address of Encrypted Wallet 1
-ENC_TEST_FIO_ADDRESS_NAME=
-
-# FIO Address of Encrypted Wallet 2
-ENC_TEST_FIO_ADDRESS_NAME_2=
-
-
-# Used For Local FIO server
-# Wallet 1
-LOCAL_PRIVATE_KEY=
-LOCAL_PUBLIC_KEY=
-
-# Wallet 2
-LOCAL_PRIVATE_KEY_2=
-LOCAL_PUBLIC_KEY_2=
-
-# FIO Address of Wallet 1
-LOCAL_TEST_FIO_ADDRESS_NAME=
-
-# FIO Address of Wallet 2
-LOCAL_TEST_FIO_ADDRESS_NAME_2=
-
-# FIO Domain of Wallet 1
-LOCAL_TEST_FIO_DOMAIN_NAME=
-
-# Encrypted Wallet 1
-LOCAL_ENC_PRIVATE_KEY=
-LOCAL_ENC_PUBLIC_KEY=
-
-# Encrypted Wallet 2
-LOCAL_ENC_PRIVATE_KEY_2=
-LOCAL_ENC_PUBLIC_KEY_2=
-
-# FIO Address of Encrypted Wallet 1
-LOCAL_ENC_TEST_FIO_ADDRESS_NAME=
-
-# FIO Address of Encrypted Wallet 2
-LOCAL_ENC_TEST_FIO_ADDRESS_NAME_2=
+# Local faucet
+LOCAL_FAUCET_PRIVATE_KEY=
+LOCAL_FAUCET_PUBLIC_KEY=
