+++ lib/transactions/queries/GetAccountPubKey.d.ts
@@ -1,4 +1,4 @@
-import { AccountResponse } from '../../entities';
+import { GetAccountPubKeyResponse } from '../../entities';
 import { RequestConfig } from '../Transactions';
 import { Query } from './Query';
 export type AccountPubKeyQueryProps = {
@@ -7,7 +7,7 @@ export type AccountPubKeyQueryProps = {
 export type AccountPubKeyQueryData = {
     account: string;
 };
-export declare class GetAccountPubKey extends Query<AccountPubKeyQueryData, AccountResponse> {
+export declare class GetAccountPubKey extends Query<AccountPubKeyQueryData, GetAccountPubKeyResponse> {
     props: AccountPubKeyQueryProps;
     ENDPOINT: "chain/get_account_fio_public_key";
     constructor(config: RequestConfig, props: AccountPubKeyQueryProps);
