+++ src/transactions/queries/GetAccountPubKey.ts
@@ -1,4 +1,4 @@
-import {AccountResponse, EndPoint} from '../../entities'
+import {GetAccountPubKeyResponse, EndPoint} from '../../entities'
 import { RequestConfig } from '../Transactions'
 import { Query } from './Query'
 
@@ -10,7 +10,7 @@ export type AccountPubKeyQueryData = {
     account: string,
 }
 
-export class GetAccountPubKey extends Query<AccountPubKeyQueryData, AccountResponse> {
+export class GetAccountPubKey extends Query<AccountPubKeyQueryData, GetAccountPubKeyResponse> {
     public ENDPOINT = `chain/${EndPoint.getAccountFioPublicKey}` as const
 
     constructor(config: RequestConfig, public props: AccountPubKeyQueryProps) {
