+++ tests/constants.ts
@@ -1,0 +1,11 @@
+import { FIOSDK } from '../src/FIOSDK';
+
+export const defaultBundledSets = 1;
+
+export const fioTokenCode = 'FIO';
+export const fioChainCode = 'FIO';
+
+export const ethTokenCode = 'ETH';
+export const ethChainCode = 'ETH';
+
+export const defaultFee = 1500 * FIOSDK.SUFUnit;
